+++ src/Popover/index.ts
@@ -1,32 +1,2 @@
-export { PopoverArrowIcon } from './ArrowIcon';
-export {
-  PopoverArrow,
-  type PopoverArrowAtomProps,
-  PopoverBackdrop,
-  PopoverPopup,
-  type PopoverPopupAtomProps,
-  PopoverPortal,
-  type PopoverPortalAtomProps,
-  PopoverPositioner,
-  type PopoverPositionerAtomProps,
-  PopoverRoot,
-  PopoverTriggerElement,
-  type PopoverTriggerElementProps,
-  PopoverViewport,
-  type PopoverViewportAtomProps,
-} from './atoms';
-export type { PopoverContextValue } from './context';
-export { PopoverProvider, usePopoverContext } from './context';
-export { default, parseTrigger } from './Popover';
-export { default as PopoverGroup } from './PopoverGroup';
-export { usePopoverPortalContainer } from './PopoverPortal';
-export type {
-  PopoverBackdropProps,
-  PopoverPlacement,
-  PopoverPopupProps,
-  PopoverPortalProps,
-  PopoverPositionerProps,
-  PopoverProps,
-  PopoverTrigger,
-  PopoverTriggerComponentProps,
-} from './type';
+export * from '@/base-ui/Popover';
+export { default } from '@/base-ui/Popover';
