+++ src/base-ui/Popover/demos/tooltip-hover-stay.tsx
