+++ src/base-ui/index.ts
@@ -8,6 +8,7 @@ export {
 } from './ContextMenu';
 export * from './DropdownMenu';
 export * from './Modal';
+export * from './Popover';
 export * from './ScrollArea';
 export { default as Select } from './Select';
 export * from './Select';
